+++ Changelog.md
@@ -6,7 +6,7 @@ All notable changes to CCP4Py will be documented in this file.
 
 - refmac.py now auto detects FP, SIGFP, and FREER labels from the input MTZ.
 - refmac.py user can now select to use F+,F- labels or HL coef. using --labels option.
-
+- refmac.py Enabled logview option.
 
 ## 0.1.0 2021-02-21